\name{do.one.experiment.t}
\alias{do.one.experiment.t}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simulation function for calculating the expected number of detectable proteins from the three stage proteomic study  
}
\description{
Simulation function for calculating the expected number of detectable proteins from the three stage proteomic study. It is used in the optim.two.stage.single().
}
\usage{
do.one.experiment.t(initial)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{initial}{
The design parameters selected at the current search
}
 }
