\name{Ftest.Ttest}
\alias{Ftest.Ttest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to perform the paired t test for each protein, and the Hotelling T test for the group that the protein is assigned. 
}
\description{
The function calculate the t test for each protein and the Hoteling T test for each group. 
}
\usage{
Ftest.Ttest(sample, n, m)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sample}{
The protein sample simulated from the protein matrix
}
  \item{n}{
The number of biological samples
}
  \item{m}{
The number of proteins
}
}
