\name{genseq.appr}
\alias{genseq.appr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A sub function to generate the sub solution space
}
\description{
A function to generate the sub solution space in the optim.two.stage.app() function
}
\usage{
genseq.appr(initial)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{initial}{
The boundary for the sub solution space
}
}
\details{
This function inputs the user defined boundary for the design parameter, the radius for the current sub-solution space. It outputs the selected solution. 
}
