\name{power.appr}
\alias{power.appr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The power function used in the optim.two.stage.appr function
}
\description{
The power function used in the optim.two.stage.appr() function. It uses the analytical approximation to estimate the expected number of detectable proteins with the true effects, and use the grouping information. 
}
\usage{
power.appr(initial)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{initial}{
The current solution selected from the solution space
}
}
