\name{Ttest}
\alias{Ttest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The function to calculate the p values for each protein in the optim.two.stage.single function
}
\description{
The function to calculate the p values for each protein in the optim.two.stage.single function. It inputs the protein sample, number of samples and number of proteins.
}
\usage{
Ttest(sample, n, m)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sample}{
The protein sample
}
  \item{n}{
The number of biological samples
}
  \item{m}{
The number of the proteins
}
}
